requirements.txt
opencv-python
pyusb
imutils
easydict
